Fix orphan instance warning
