+++ kore/test/Test/Kore/Builtin/Map.hs
@@ -24,6 +24,7 @@ import           Kore.Attribute.Hook
 import qualified Kore.Attribute.Symbol as StepperAttributes
 import qualified Kore.Builtin.AssociativeCommutative as Ac
 import qualified Kore.Builtin.Map as Map
+import qualified Kore.Builtin.MapSymbols as Map
 import qualified Kore.Domain.Builtin as Domain
 import           Kore.IndexedModule.MetadataTools
                  ( SmtMetadataTools )
